bq + sagemaker renames
